gaybuster_service/pyproject.toml
[tool.poetry]
name = "gaybuster_service"
version = "0.1.0"
description = ""
authors = ["yilbegan <w2v@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
tensorflow = "^2.2.0"
opencv-python = "^4.2.0"
dlib = "^19.20.0"
environs = "^8.0.0"
Pillow = "^7.1.2"
loguru = "^0.5.0"
aio-pika = "^6.6.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
